// packages/@aws-cdk/aws-cloudwatch/lib/graph.d.ts
export interface HorizontalAnnotation {
    /**
     * The value of the annotation
     */
    readonly value: number;
    /**
     * Label for the annotation
     *
     * @default - No label
     */
    readonly label?: string;
    /**
     * The hex color code, prefixed with '#' (e.g. '#00ff00'), to be used for the annotation.
     * The `Color` class has a set of standard colors that can be used here.
     *
     * @default - Automatic color
     */
    readonly color?: string;
    /**
     * Add shading above or below the annotation
     *
     * @default No shading
     */
    readonly fill?: Shading;
    /**
     * Whether the annotation is visible
     *
     * @default true
     */
    readonly visible?: boolean;
}
/**
 * Fill shading options that will be used with an annotation
 */
export declare enum Shading {
    /**
     * Don't add shading
     */
    NONE = "none",
    /**
     * Add shading above the annotation
     */
    ABOVE = "above",
    /**
     * Add shading below the annotation
     */
    BELOW = "below"
}